Refactor acc to dom's structure
